+++ lib/s_universal_html/src/html.dart
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+@Deprecated(
+    'Internal library. Import package:s_packages/s_universal_html/html.dart instead.')
+// ignore: invalid_internal_annotation
+@internal
 library;
 
 import 'dart:async';